:recycle: [#1090] Refactor tests
