+++ src/openforms/forms/tests/test_data_migrations.py
@@ -1,44 +1,7 @@
-from django.db import connection
-from django.db.migrations.executor import MigrationExecutor
-from django.test import TestCase
+from openforms.utils.tests.test_migrations import TestMigrations
 
 
-class TestMigrations(TestCase):
-    """
-    Adapted from https://www.caktusgroup.com/blog/2016/02/02/writing-unit-tests-django-migrations/
-    """
-
-    app = None
-    migrate_from = None
-    migrate_to = None
-
-    def setUp(self):
-        assert self.migrate_from and self.migrate_to and self.app, (
-            "TestCase '%s' must define migrate_from, migrate_to and app properties"
-            % type(self).__name__
-        )
-        self.migrate_from = [(self.app, self.migrate_from)]
-        self.migrate_to = [(self.app, self.migrate_to)]
-        executor = MigrationExecutor(connection)
-        old_apps = executor.loader.project_state(self.migrate_from).apps
-
-        # Reverse to the original migration
-        executor.migrate(self.migrate_from)
-
-        self.setUpBeforeMigration(old_apps)
-
-        # Run the migration to test
-        executor = MigrationExecutor(connection)
-        executor.loader.build_graph()  # reload.
-        executor.migrate(self.migrate_to)
-
-        self.apps = executor.loader.project_state(self.migrate_to).apps
-
-    def setUpBeforeMigration(self, apps):
-        pass
-
-
-class ReplateWidgetDataMigrationTests(TestMigrations):
+class ReplaceWidgetDataMigrationTests(TestMigrations):
     migrate_from = "0012_merge_20211222_0831"
     migrate_to = "0013_auto_20211220_1755"
     app = "forms"
